+++ R/tidy_add_reference_rows.R
@@ -72,14 +72,9 @@ tidy_add_reference_rows <- function(
   if (inherits(model, "aov")) {
     return(x %>% dplyr::mutate(reference_row = NA))
   }
-  # adding reference rows is not meaningful for marginal effects
-  if (isTRUE(.attributes$coefficients_type == "marginal_effects")) {
+  # checking cases where adding reference rows is not meaningful
+  if (isTRUE(.attributes$skip_add_reference_rows))
     return(x %>% dplyr::mutate(reference_row = NA))
-  }
-  # adding reference rows is not meaningful for conditional effects
-  if (isTRUE(.attributes$coefficients_type == "conditional_effects")) {
-    return(x %>% dplyr::mutate(reference_row = NA))
-  }
 
   if ("header_row" %in% names(x)) {
     stop("`tidy_add_reference_rows()` cannot be applied after `tidy_add_header_rows().`")